Trace and timestamp all user-entered commands
In addition to just running `set -x` as the last command in bashrc,
there's also some funny stuff going on with `set +x` invocations in PS1.
We do this to prevent command tracing from becoming a firehose of noise.

With *just* `set -x` in an interactive non-login shell, sub-shells
spawned while expanding PS1 will also be traced, resulting in PS1
effectively becoming the following:

```
++ date --iso-8601=seconds
++ :
++ __git_ps1
++ local exit=0
++ local pcmode=no
++ local detached=no
++ local 'ps1pc_start=\u@\h:\w '
++ local 'ps1pc_end=\$ '
++ local 'printf_format= (%s)'
++ case "$#" in
++ printf_format=' (%s)'
++ local ps1_expanded=yes
++ '[' -z '' ']'
++ '[' -z '5.0.16(1)-release' ']'
++ shopt -q promptvars
++ local repo_info rev_parse_exit_code
+++ git rev-parse --git-dir --is-inside-git-dir --is-bare-repository --is-inside-work-tree --short HEAD
++ repo_info=
++ rev_parse_exit_code=128
++ '[' -z '' ']'
++ return 0
++ exit_code=0
+++ '[' 0 == 0 ']'
+++ echo 33m
++ color_code='01;33m'
++ echo '[0]'
sxlijin@sxlijin-debian:2020-04-26T21:00:26-07:00:~ [0]
```

This effectively renders the shell unusable. Unfortunately, it doesn't
seem like there's any way to suppress tracing during PS1 expansion
itself (at least, not if we intend for this to work on any bash), so
instead, we simply disable tracing in every instance of command
substitution in PS1, which improves things substantially:

```
++ set +x
++ set +x
++ set +x
sxlijin@sxlijin-debian:2020-04-26T21:12:51-07:00:~ [0]
```

But those last few `set +x` traces are still annoying, but since we
still need to run them, we black-hole their output using a command group
(note that we use {} for the command group because we need the `set +x`
to execute in the top-level shell context of the command substitution,
not a subshell like () would do).

Lastly, we also modify PS4 to dump the timestamp of command execution.

----

The inspiration for this comes from @countrmeasure, via this [blog
post](https://redandblack.io/blog/2020/bash-prompt-with-updating-time/).
Prior to that I didn't even know there were other PS* parameters. I
really like the idea of recording when a command is actually executed;
it's the reason I have a `$(date)` in my PS1, except unfortunately as
that blog post notes, this means that the timestamp is recorded when the
prompt is written, aka when the previous command completes.

I didn't like the approach of cursor-fiddling though, so I dug a little
deeper, and started with the documentation for PS* (a.k.a. `man bash` -
incidentally, also the first time I've ever looked at that man page).
I started by fooling around a bit with PS0 (I originally was thinking that
since PS1 gives me timing information about command *completion*, I just
need to add timing information about command *start*, which is exactly
what PS0 provides a hook for):

```
sxlijin@sxlijin-debian:2020-04-26T21:42:08-07:00:~ [0]
$ PS0='$(date --iso-8601=seconds)\n'
sxlijin@sxlijin-debian:2020-04-26T21:42:13-07:00:~ [0]
$ echo thing-to-be-echoed
2020-04-26T21:42:14-07:00
thing-to-be-echoed
sxlijin@sxlijin-debian:2020-04-26T21:42:14-07:00:~ [0]
$ PS0="$PS1"
2020-04-26T21:42:24-07:00
sxlijin@sxlijin-debian:2020-04-26T21:42:24-07:00:~ [0]
$ echo thing-to-be-echoed
sxlijin@sxlijin-debian:2020-04-26T21:42:26-07:00:~ [0]
$ thing-to-be-echoed
```

Obviously, both PS0 options are pretty misleading: the first suggests
that a command is doing something it isn't and the second suggests that
the command does nothing and that its output is actually the next
command. At this point I realized that to dump a timestmap right before
the command executes, you not only need to dump the timestamp, but you
also need to dump the command again, and hey, that sounds *remarkably*
like command tracing, and it so happens that I just learned that PS4
controls the command tracing output prefix... and this commit is the
rest of the story.

Incidentally, I've actually fooled around with the idea of `set -x`ing
my prompt before, but ran into the noisy PS1 tracing and quickly gave
up. Probably didn't want to bother with any more yakshaving at that
point... :)