tmux: stay in copy mode after yanking
I've sometimes found that when scrolling up to view some old terminal
output, there's more than one thing I want to copy out. Today that
workflow requires scrolling all the way back up to find the same
position in terminal output for each thing I want to copy out; this will
make it so that I don't have to scroll back up.

The drawback of this is that in situations where I only want to copy one
thing, I'll have to [Enter] after yanking the highlighted text to exit
copy mode and return to the current terminal context, but I don't expect
this to be a particularly problematic workflow.