vimrc: always show a status line
laststatus controls whether or not a status message is shown at the
bottom of a Vim window; it defaults to laststatus=1, which means that
the status is only shown if there are multiple tabs. Setting
laststatus=2 means that the status will always be shown, regardless of
whether or not multiple tabs are open.

This comes in handy when, for example, I want to quickly see the current
filename and don't want to look up the command that will do that for me.