vimrc: auto-indent new comments reasonably
Turning smartindent on results in

  1.  forcing new comments (at least, those delimited by #) to start from
      column 0, ignoring whatever indentation the cursor was at prior to
      entering the #, and

  2.  rendering >> a noop on comments that have a # in column 0.

This is very undesirable behavior - it's caused me no end of frustration
over the years - and recent StackExchange answers suggest that
smartindent is *very* outdated nowadays and that `filetype indent on` is
more than sufficient.

There are some other answers that suggest using `cindent` and `cinkeys`
(see https://unix.stackexchange.com/a/543571), but `set cindent
cinkeys-=#0` screws with automatic indentation in Git commit messages
(and presumably other plaintext files) so since the file-aware
indentation is sufficient we don't bother with it.