bashrc: fix exit code in PS1
Somehow apparently the exit code in my PS1 has been broken since I
originally checked this in, much to my surprise: it was actually
previously showing the exit code of the `date` invocation in the PS1.

I've now replaced the PS1 completely with a command substitution
sequence of echoes which makes all of that much saner and now I think
also easier to read (although admittedly slightly more frustrating,
since it's a sequence of echo commands now).