bashrc: handle unset PROMPT_COMMAND
Leave PROMPT_COMMAND unset if it was not set, since Bash does not
consider consecutive ;s to be a valid command and complains.

This also means that fixing PROMPT_COMMAND to disable command tracing
inside it is cleaner: we only have to strip a single trailing ;, as
opposed to arbitrarily many.