fix(vim): better shortcuts for file nav
Using `:split +Explore` and `:vsplit +Explore` are bad choices, because
the default behavior of `+Explore` is to open in a _new_ split if the
current buffer has not been saved since its last modification, which
makes sense: the command has no way of guaranteeing you're not losing
track of a modified buffer.

There are commands that exist for this (which I didn't know about):
namely, `Sexplore` for horizontal splitting and `Sexplore!` for vertical
splitting. (These commands respect your split config: `Hexplore[!]` and
`Vexplore[!]` do not.) To make them easy to use (particularly vertical
splitting, since ! is a bit of a reach with my hand), map `s` and `v` to
them (which do not actually map to any commands).