fix(xsessionrc): use gsettings if on XWayland
Starting in Jammy 22.04 [1], Wayland, not X11, is now the default
display server.  This has the unfortunate consequence that `setxkbmap`,
which we use to remap `CapsLock` to `Ctrl`, no longer works.

When running Wayland, there's still an X server, but it's provided by
XWayland, an emulation layer. This emulation layer will result in
`setxkbmap` invocations silently failing. The fix is to use GNOME
settings to apply the keyboard remapping.

[1] https://ubuntu.com/blog/ubuntu-22-04-lts-whats-new-linux-desktop